requirements.txt
PyYAML>=6.0
requests>=2.31.0
openai>=1.40.0
psutil>=5.9.8
pandas>=2.2.0
statsmodels>=0.14.0
matplotlib>=3.8.0
